+++ ingest2/operations/create_manage_limits_request.go
@@ -7,6 +7,7 @@ import (
 )
 
 type createManageLimitsRequestOpHandler struct {
+	effectsProvider
 }
 
 // Details returns details about create limits request operation
@@ -24,11 +25,4 @@ func (h *createManageLimitsRequestOpHandler) Details(op rawOperation,
 			RequestID: int64(opRes.MustCreateManageLimitsRequestResult().MustSuccess().ManageLimitsRequestId),
 		},
 	}, nil
-}
-
-//ParticipantsEffects returns source participant effect
-func (h *createManageLimitsRequestOpHandler) ParticipantsEffects(opBody xdr.OperationBody,
-	opRes xdr.OperationResultTr, source history2.ParticipantEffect, _ []xdr.LedgerEntryChange,
-) ([]history2.ParticipantEffect, error) {
-	return []history2.ParticipantEffect{source}, nil
 }