+++ ingest2/operations/create_pre_issuance_request.go
@@ -7,6 +7,7 @@ import (
 )
 
 type createPreIssuanceRequestOpHandler struct {
+	effectsProvider
 }
 
 // Details returns details about create pre issuance request operation
@@ -25,11 +26,4 @@ func (h *createPreIssuanceRequestOpHandler) Details(op rawOperation,
 			IsFulfilled: successResult.Fulfilled,
 		},
 	}, nil
-}
-
-//ParticipantsEffects returns source participant effect
-func (h *createPreIssuanceRequestOpHandler) ParticipantsEffects(opBody xdr.OperationBody,
-	opRes xdr.OperationResultTr, source history2.ParticipantEffect, _ []xdr.LedgerEntryChange,
-) ([]history2.ParticipantEffect, error) {
-	return []history2.ParticipantEffect{source}, nil
 }