+++ ingest2/operations/manage_key_value.go
@@ -9,6 +9,7 @@ import (
 )
 
 type manageKeyValueOpHandler struct {
+	effectsProvider
 }
 
 // Details returns details about manage key value operation
@@ -48,11 +49,4 @@ func (h *manageKeyValueOpHandler) Details(op rawOperation, _ xdr.OperationResult
 			Value:  value,
 		},
 	}, nil
-}
-
-//ParticipantsEffects - returns source of the operation
-func (h *manageKeyValueOpHandler) ParticipantsEffects(opBody xdr.OperationBody,
-	_ xdr.OperationResultTr, source history2.ParticipantEffect, _ []xdr.LedgerEntryChange,
-) ([]history2.ParticipantEffect, error) {
-	return []history2.ParticipantEffect{source}, nil
 }