+++ ingest2/operations/review_request_withdraw.go
@@ -1,40 +1,41 @@
 package operations
 
 import (
-	"gitlab.com/tokend/go/xdr"
 	"gitlab.com/tokend/horizon/db2/history2"
 	"gitlab.com/tokend/horizon/ingest2/internal"
 	"gitlab.com/tokend/regources/v2"
 )
 
 type withdrawHandler struct {
-	balanceProvider balanceProvider
+	effectsProvider
 }
 
-//ParticipantsEffects - returns effect due to review of withdrawal request
-func (h *withdrawHandler) ParticipantsEffects(op xdr.ReviewRequestOp,
-	res xdr.ExtendedResult, request xdr.ReviewableRequestEntry,
-	source history2.ParticipantEffect, ledgerChanges []xdr.LedgerEntryChange,
-) ([]history2.ParticipantEffect, error) {
-	details := request.Body.MustWithdrawalRequest()
+//Fulfilled - returns slice of effects for participants of the operation
+func (h *withdrawHandler) Fulfilled(details requestDetails) ([]history2.ParticipantEffect, error) {
+	request := details.Request.Body.MustWithdrawalRequest()
 
 	effect := history2.Effect{
-		Type: history2.EffectTypeChargedFromLocked,
-		ChargedFromLocked: &history2.BalanceChangeEffect{
-			Amount: regources.Amount(details.Amount),
-			Fee:    internal.FeeFromXdr(details.Fee),
+		Type: history2.EffectTypeWithdrawn,
+		Withdrawn: &history2.BalanceChangeEffect{
+			Amount: regources.Amount(request.Amount),
+			Fee:    internal.FeeFromXdr(request.Fee),
 		},
 	}
 
-	if op.Action != xdr.ReviewRequestOpActionApprove {
-		effect = history2.Effect{
-			Type: history2.EffectTypeWithdrawn,
-			Withdrawn: &history2.BalanceChangeEffect{
-				Amount: regources.Amount(details.Amount),
-				Fee:    internal.FeeFromXdr(details.Fee),
-			},
-		}
+	return h.BalanceEffectWithAccount(details.SourceAccountID, request.Balance, &effect), nil
+}
+
+//PermanentReject - returns participants of fully rejected request
+func (h *withdrawHandler) PermanentReject(details requestDetails) ([]history2.ParticipantEffect, error) {
+	request := details.Request.Body.MustWithdrawalRequest()
+
+	effect := history2.Effect{
+		Type: history2.EffectTypeUnlocked,
+		Unlocked: &history2.BalanceChangeEffect{
+			Amount: regources.Amount(request.Amount),
+			Fee:    internal.FeeFromXdr(request.Fee),
+		},
 	}
 
-	return populateEffects(h.balanceProvider.MustBalance(details.Balance), effect, source), nil
+	return h.effectsProvider.BalanceEffectWithAccount(details.SourceAccountID, request.Balance, &effect), nil
 }